Refactor signOrder out and fix lint
