+++ src/contracts/prices/calls/getPrice.ts
@@ -10,7 +10,6 @@ export const getPrice = async (
   environment: Environment,
   contractAddress: string,
   token: TokenInterface,
-  preventCancelDown: boolean = false,
 ): Promise<PriceInterface> => {
   const quoteToken = await getQuoteToken(environment, contractAddress);
   const contract = await getContract(
@@ -22,5 +21,5 @@ export const getPrice = async (
   const { 0: price } = await contract.methods.getPrice(token.address).call();
   const base = createQuantity(token, appendDecimals(token, 1));
   const quote = createQuantity(quoteToken, price);
-  return createPrice(base, quote, preventCancelDown);
+  return createPrice(base, quote);
 };
